fix(schemas): keep RepairEvidenceV1 out of substrate import chain
Move RepairEvidenceV1 to orion/schemas/repair_evidence.py so registry and
pre_turn_appraisal imports do not load orion.substrate (requests dep).
Substrate models re-export for backward compat; add import-light gate test.
